MindAIde/requirements.txt
streamlit==1.42.0
langchain
langchain_community
openai
transformers
torch
python-dotenv
langchain-openai